requirements.txt
discord.py
wikipedia
Pillow
pytz
requests
aiofiles
rapidfuzz
Wavelink
asyncpg
levenshtein
spotipy